backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.0
pillow
numpy
python-multipart
typing-extensions
opencv-python
python-dotenv
requests
huggingface-hub
pydantic
PyMuPDF

